User cache is now initiated at the global level
